third_party/blink/web_tests/fast/dom/StyleSheet/empty-shadow-style.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<div id='sandbox'></div>
<pre id='console'></pre>
<script>
var sandbox = document.getElementById('sandbox');
var host = document.createElement('div');
var shadowRoot = host.attachShadow({mode: 'open'});
shadowRoot.innerHTML = '<style></style><div>Shadow</div>';
shadowRoot.appendChild(document.createElement('style'));
sandbox.appendChild(host);

shouldBe("shadowRoot.styleSheets.length", "2");
</script>
</body>
</html>
