third_party/blink/web_tests/fast/dom/StyleSheet/empty-shadow-style-expected.txt

PASS shadowRoot.styleSheets.length is 2
PASS successfullyParsed is true

TEST COMPLETE
